OK-512: Refaktoroitu harkinnanvaraiset-tabin koodia pienemmiksi paloiksi
