+++ src/app/haku/[oid]/hakukohde/[hakukohde]/harkinnanvaraiset/components/harkinnanvaraiset-table.tsx
@@ -71,7 +71,6 @@ export const HarkinnanvaraisetTable = ({
               }}
               clearable={true}
               placeholder="(ei valintaa)"
-              name={`${props.hakemusOid}_harkinnanvarainenTila`}
               options={[
                 {
                   label: 'Hyväksytty',
